+++ docker-images/kafka-connect/scripts/kafka_connect_config_generator.sh
@@ -18,6 +18,23 @@ consumer.ssl.truststore.location=/tmp/kafka/cluster.truststore.p12
 consumer.ssl.truststore.password=${CERTS_STORE_PASSWORD}
 EOF
 )
+
+    if [ -n "$KAFKA_CONNECT_TLS_AUTH_CERT" ] && [ -n "$KAFKA_CONNECT_TLS_AUTH_KEY" ]; then
+        TLS_AUTH_CONFIGURATION=$(cat <<EOF
+ssl.keystore.location=/tmp/kafka/cluster.keystore.p12
+ssl.keystore.password=${CERTS_STORE_PASSWORD}
+ssl.keystore.type=PKCS12
+
+producer.ssl.keystore.location=/tmp/kafka/cluster.keystore.p12
+producer.ssl.keystore.password=${CERTS_STORE_PASSWORD}
+producer.ssl.keystore.type=PKCS12
+
+consumer.ssl.keystore.location=/tmp/kafka/cluster.keystore.p12
+consumer.ssl.keystore.password=${CERTS_STORE_PASSWORD}
+consumer.ssl.keystore.type=PKCS12
+EOF
+)
+    fi
 fi
 
 # Write the config file
@@ -34,4 +51,5 @@ plugin.path=${KAFKA_CONNECT_PLUGIN_PATH}
 ${KAFKA_CONNECT_CONFIGURATION}
 
 ${TLS_CONFIGURATION}
+${TLS_AUTH_CONFIGURATION}
 EOF